Unify some configs (#277)
