+++ crates/iceberg-ext/src/configs/primitives/mod.rs
@@ -1,0 +1,2 @@
+mod location;
+pub use location::Location;
